# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Wakarimi Music Files API",
        "contact": {
            "name": "Dmitry Kolesnikov (Zalimannard)",
            "email": "zalimannard@mail.ru"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.4"
    },
    "host": "localhost:8022",
    "basePath": "/api",
    "paths": {
...
        "/dirs/scan": {
            "post": {
                "description": "Initiates a scan in all directories to identify new or updated files.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Directories"
                ],
                "summary": "Scan all directories",
                "responses": {
                    "200": {
                        "description": "All directories scanned successfully"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/dirs/{dirId}/content": {
            "get": {
                "description": "Retrieves a list of subdirectories for a given directory ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Directories"
                ],
                "summary": "Retrieve content of a directory by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Directory ID",
                        "name": "dirId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dir_handler.contentResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid dirId format",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "404": {
                        "description": "Directory not found",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/dirs/{dirId}/scan": {
            "post": {
                "description": "Initiates a scan in the specified directory to identify new or updated files.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Directories"
                ],
                "summary": "Scan a directory by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Directory ID",
                        "name": "dirId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Directory scanned successfully"
                    },
                    "400": {
                        "description": "Invalid dirId format",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "404": {
                        "description": "Directory not found",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
...
        "/songs/{songId}/cover": {
            "get": {
                "description": "Retrieves detailed information about a cover for a song by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Covers"
                ],
                "summary": "Retrieve a cover for a song by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Song ID",
                        "name": "songId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/song_handler.getCoverResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid songId format",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "404": {
                        "description": "Cover not found",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/songs/{songId}/download": {
            "get": {
                "description": "Downloads a song file identified by the songId",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Download a song by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Song ID",
                        "name": "songId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Song File",
                        "schema": {
                            "type": "file"
                        },
                        "headers": {
                            "Content-Disposition": {
                                "type": "string",
                                "description": "attachment; filename=[name of the file]"
                            },
                            "Content-Type": {
                                "type": "string",
                                "description": "application/octet-stream"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid songId format",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error, Failed to calculate absolute path",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dir_handler.contentResponse": {
            "type": "object",
            "properties": {
                "dirs": {
                    "description": "Array containing directory items",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dir_handler.contentResponseDirItem"
                    }
                },
                "songs": {
                    "description": "Array containing song items",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dir_handler.contentResponseSongItem"
                    }
                }
            }
        },
        "dir_handler.contentResponseDirItem": {
            "type": "object",
            "properties": {
                "dirId": {
                    "description": "Unique identifier for the directory",
                    "type": "integer"
                },
                "name": {
                    "description": "Name of the directory",
                    "type": "string"
                }
            }
        },
        "dir_handler.contentResponseSongItem": {
            "type": "object",
            "properties": {
                "bitrateKbps": {
                    "description": "Bitrate in kilobits per second",
                    "type": "integer"
                },
                "channelsN": {
                    "description": "Number of audio channels",
                    "type": "integer"
                },
                "dirId": {
                    "description": "Directory identifier where the song resides",
                    "type": "integer"
                },
                "durationMs": {
                    "description": "Duration of the song in milliseconds",
                    "type": "integer"
                },
                "extension": {
                    "description": "File extension of the song",
                    "type": "string"
                },
                "filename": {
                    "description": "Filename of the song",
                    "type": "string"
                },
                "lastContentUpdate": {
                    "description": "Time of the last update to the song's content",
                    "type": "string"
                },
                "sampleRateHz": {
                    "description": "Sample rate in hertz",
                    "type": "integer"
                },
                "sha256": {
                    "description": "SHA-256 hash of the file",
                    "type": "string"
                },
                "sizeByte": {
                    "description": "File size in bytes",
                    "type": "integer"
                },
                "songId": {
                    "description": "Unique identifier for the song",
                    "type": "integer"
                }
            }
        },
...
        "responses.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "song_handler.getCoverResponse": {
            "type": "object",
            "properties": {
                "coverId": {
                    "description": "Unique identifier for the cover.",
                    "type": "integer"
                },
                "extension": {
                    "description": "File extension of the cover.",
                    "type": "string"
                },
                "heightPx": {
                    "description": "Height of the cover in pixels.",
                    "type": "integer"
                },
                "lastContentUpdate": {
                    "description": "Timestamp of the last content update.",
                    "type": "string"
                },
                "sha256": {
                    "description": "SHA-256 hash of the cover.",
                    "type": "string"
                },
                "sizeByte": {
                    "description": "File size of the cover in bytes.",
                    "type": "integer"
                },
                "widthPx": {
                    "description": "Width of the cover in pixels.",
                    "type": "integer"
                }
            }
        },
...
    }
}